RNG doc
